+++ src/pages/auth/signup/Signup.tsx
@@ -2,18 +2,29 @@ import * as styles from './Signup.css';
 import Button from '@shared/components/button/Button';
 import SignupForm from './components/SignupForm';
 import { useSignupForm, type SignupFormData } from './hooks/useSignupForm';
+import { useSignupMutations } from './hooks/useSignupMutations';
 
 export default function Signup() {
   const {
     form,
     isEmailSent,
+    setIsEmailSent,
     emailError,
+    setEmailError,
     isEmailVerified,
+    setIsEmailVerified,
+    verifiedEmail,
+    setVerifiedEmail,
     codeError,
-    requestEmailVerification,
-    verifyCode,
-    submitSignup,
+    setCodeError,
+    resetEmailVerificationState,
   } = useSignupForm();
+
+  const {
+    signupMutation,
+    verificationConfirmMutation,
+    verificationRequestMutation,
+  } = useSignupMutations();
 
   const { handleSubmit, formState: { errors, isValid }, watch, setValue } = form;
 
@@ -64,22 +75,88 @@ export default function Signup() {
     setValue('phone', e.target.value, { shouldValidate: true });
   };
 
+  const getHttpStatus = (error: unknown): number | undefined => {
+    if (error && typeof error === 'object' && 'response' in error &&
+        error.response && typeof error.response === 'object' && 'status' in error.response) {
+      return error.response.status as number;
+    }
+    return undefined;
+  };
+
   const handleEmailVerification = async () => {
     if (!emailValue) return;
-    await requestEmailVerification(emailValue);
+    
+    try {
+      resetEmailVerificationState();
+      await verificationRequestMutation.mutateAsync({ email: emailValue });
+      setIsEmailSent(true);
+    } catch (error: unknown) {
+      console.error('이메일 인증 요청 실패');
+      setIsEmailSent(false);
+      
+      // 409 상태 코드인 경우 이미 존재하는 이메일
+      const status = getHttpStatus(error);
+      if (status === 409) {
+        setEmailError('이미 존재하는 이메일 입니다.');
+      } else {
+        setEmailError('이메일 인증 요청에 실패했습니다.');
+      }
+    }
   };
 
   const handleCodeVerification = async () => {
     if (!verificationCodeValue) return;
-    await verifyCode(verificationCodeValue);
+    
+    try {
+      setCodeError('');
+      const currentEmail = form.getValues('email');
+      if (!currentEmail) {
+        setCodeError('이메일을 먼저 입력해주세요.');
+        return;
+      }
+      
+      await verificationConfirmMutation.mutateAsync({ 
+        email: currentEmail, 
+        code: verificationCodeValue 
+      });
+      setIsEmailVerified(true);
+      setVerifiedEmail(currentEmail);
+    } catch (error: unknown) {
+      console.error('인증번호 확인 실패');
+      
+      // 400 상태 코드인 경우 인증번호 불일치
+      const status = getHttpStatus(error);
+      if (status === 400) {
+        setCodeError('인증번호가 일치하지 않습니다.');
+      } else {
+        setCodeError('잘못된 인증번호입니다.');
+      }
+    }
   };
 
   const onSubmit = async (data: SignupFormData) => {
-    const result = await submitSignup(data);
-    if (result.success) {
+    try {
+      if (!isEmailVerified) {
+        alert('이메일 인증을 완료해주세요.');
+        return;
+      }
+      if (data.email !== verifiedEmail) {
+        alert('인증된 이메일과 입력한 이메일이 일치하지 않습니다.');
+        return;
+      }
+      
+      await signupMutation.mutateAsync({
+        email: data.email,
+        name: data.name,
+        password: data.password,
+        studentNumber: data.studentNumber,
+        phone: data.phone,
+      });
+      
       console.log('회원가입 성공');
-    } else {
-      console.error(result.error);
+    } catch (error) {
+      console.error('회원가입 실패:', error);
+      alert('회원가입에 실패했습니다.');
     }
   };
